febio2: cleaned up compiler warnings by including 'override' commands for functions derived from virtual base class, including formatting strings in printf where needed, and defining virtual destructors where needed.
git-svn-id: https://gforge.sci.utah.edu/svn/MRLProjects/FEBio2@10172 b505bdfe-9d5c-0410-8600-697d5fc8e487

+++ FEBioMech/FELinearSolidSolver.cpp
@@ -82,7 +82,7 @@ bool FELinearSolidSolver::Quasin(double time)
 
 	// start Newton-loop
 	bool bconv = false;
-	int N = m_u.size();
+	int N = (int)m_u.size();
 	vector<double> du(N), Du(N), U(m_u), D(m_d); zero(du); zero(Du);
 	const int NMAX = 10;
 	int n = 0;
